requirements.txt
# ============================================================================
# Airfoil RL Optimizer - Requirements
# ============================================================================
# Production-grade reinforcement learning for aerodynamic optimization
# Author: Mohamed Noorul Naseem

# Core Scientific Computing
numpy>=1.21.0
scipy>=1.9.0
matplotlib>=3.5.0

# Machine Learning & RL
stable-baselines3>=2.0.0
gymnasium>=0.29.0
torch>=2.0.0

# Web Application
# Web Application (Dash)
dash>=2.14.0
dash-bootstrap-components>=1.5.0
pandas>=2.0.0
plotly>=5.18.0

# Configuration
PyYAML>=6.0

# ============================================================================
# External Tools (install separately):
# ============================================================================
# 
# XFOIL (MIT Panel Method CFD):
#   Windows: Download from https://web.mit.edu/drela/Public/web/xfoil/
#   Linux:   sudo apt-get install xfoil
#   macOS:   brew install xfoil
#
# Stanford SU2 (Advanced CFD):
#   https://su2code.github.io/download.html
# ============================================================================
